Initial Release
This is the first release of Vote for Constellation.  This initial
release packs in a lot of cool features so here is a brief overview:

  * Multiple Voting Systems
    * Plurality (First Past The Post)
    * Instant Run-Off
    * Plurality at Large (Multiple winners)
    * STV (Single Transferable Vote)
  * Elections managed by different groups
  * Private election results

As with all other Constellation Apps, this app requires the
Constellation Base component to be loaded to work correctly.  At least
version 0.1.4 is needed for full functionality.