Release v0.1.2: More Time, Withdrawals, and Vote Counts
This release includes a fix for editing the open on and close on times
for the poll as well as a fix to ballots that contain candidates that
have dropped out of the race.  An additional feature that ships with
this release is the poll results page now shows the number of votes
cast.